feat: add webpack and eslint settings to project, rename files to .ts
