+++ .eslintignore
@@ -1,0 +1,4 @@
+webpack.config.js
+webpack.dev.config.js
+webpack.prod.config.js
+node_modules
