+++ .eslintrc.json
@@ -1,21 +1,43 @@
 {
-    "plugins": ["prettier", "import"],
-    "extends": [
-        "plugin:prettier/recommended",
-        "prettier"
-    ],
-    "parserOptions": {
-        "ecmaVersion": 2020,
-        "sourceType": "module"
-    },
     "env": {
         "es6": true,
         "browser": true,
         "node": true
     },
+    "extends": [
+        "plugin:prettier/recommended",
+        "prettier",
+        "airbnb-typescript/base",
+        "eslint:recommended",
+        "plugin:@typescript-eslint/recommended"
+    ],
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "ecmaVersion": 2020,
+        "sourceType": "module",
+        "project": "./tsconfig.json"
+    },
+    "plugins": [
+        "prettier",
+        "import",
+        "@typescript-eslint"
+        ],
     "rules": {
+        // "no-plusplus":"off",
+        "@typescript-eslint/no-explicit-any": "error",
         "no-debugger": "off",
         "no-console": 0,
-        "class-methods-use-this": "off"
+        "class-methods-use-this": "off",
+        "prettier/prettier": [
+            "error",
+            {
+                "trailingComma": "es5",
+                "endOfLine": "auto",
+                "tabWidth": 2,
+                "semi": true,
+                "singleQuote": true,
+                "printWidth": 120
+            }
+          ]
     }
 }
