+++ src/components/controller/loader.ts
@@ -1,0 +1,52 @@
+class Loader {
+  baseLink: string;
+
+  options: [];
+
+  constructor(baseLink: string, options: []) {
+    this.baseLink = baseLink;
+    this.options = options;
+  }
+
+  getResp(
+    { endpoint, options }: { endpoint: string; options?: Record<string, never>}
+    // { endpoint, options = {} },
+    callback: <T>(data: T) => void = () => {
+      console.error('No callback for GET response');
+    }
+    // type Callback<T> = (data: T) => void;
+  ) {
+    this.load('GET', endpoint, callback, options);
+  }
+
+  errorHandler(res) {
+    if (!res.ok) {
+      if (res.status === 401 || res.status === 404)
+        console.log(`Sorry, but there is ${res.status} error: ${res.statusText}`);
+      throw Error(res.statusText);
+    }
+
+    return res;
+  }
+
+  makeUrl(options, endpoint) {
+    const urlOptions = { ...this.options, ...options };
+    let url = `${this.baseLink}${endpoint}?`;
+
+    Object.keys(urlOptions).forEach((key) => {
+      url += `${key}=${urlOptions[key]}&`;
+    });
+
+    return url.slice(0, -1);
+  }
+
+  load(method, endpoint, callback, options = {}) {
+    fetch(this.makeUrl(options, endpoint), { method })
+      .then(this.errorHandler)
+      .then((res) => res.json())
+      .then((data) => callback(data))
+      .catch((err) => console.error(err));
+  }
+}
+
+export default Loader;
