+++ webpack.config.js
@@ -2,9 +2,12 @@ const path = require('path');
 const { merge } = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const ESLintPlugin = require('eslint-webpack-plugin');
 
+
+// const eSLintPlugin = (mode) =>  mode === 'dev' ? [] : [new ESLintPlugin ({ extensions: [".ts"] }) ];
 const baseConfig = {
-    entry: path.resolve(__dirname, './src/index.js'),
+    entry: path.resolve(__dirname, './src/index.ts'),
     mode: 'development',
     module: {
         rules: [
@@ -15,11 +18,11 @@ const baseConfig = {
         ],
     },
     resolve: {
-        extensions: ['.js'],
+        extensions: ['.ts', '.js'],
     },
     output: {
         filename: 'index.js',
-        path: path.resolve(__dirname, '../dist'),
+        path: path.resolve(__dirname, 'dist'),
     },
     plugins: [
         new HtmlWebpackPlugin({
@@ -27,6 +30,9 @@ const baseConfig = {
             filename: 'index.html',
         }),
         new CleanWebpackPlugin(),
+        new ESLintPlugin ({ extensions: [".ts"] })
+        // ...((mode === 'prod') ? [] : [new ESLintPlugin ({ extensions: [".ts"] }) ])
+        // ...eSLintPlugin(mode)
     ],
 };
