+++ webpack.dev.config.js
@@ -4,6 +4,6 @@ module.exports = {
     mode: 'development',
     devtool: 'inline-source-map',
     devServer: {
-        contentBase: path.resolve(__dirname, '../dist'),
+        contentBase: path.resolve(__dirname, 'dist'),
     },
 };
